+++ src/main/java/patio/voting/repositories/VoteRepository.java
@@ -37,7 +37,7 @@ public interface VoteRepository extends PageableRepository<Vote, UUID> {
    * @param voting the voting this vote belongs to
    * @return the avg score value
    */
-  Integer findAvgScoreByVoting(Voting voting);
+  Float findAvgScoreByVoting(Voting voting);
 
   /**
    * Finds a vote created by some {@link User} in some {@link Voting}
